Add script version 2.0
During the first version of the script, it was detected that the uninstallation of the Teamviewer application was not executed on some computers. The logs showed that the application could not be uninstalled since the service was running. To stop the teamviewer service it is necessary for the script to have administrator permissions, for which it was necessary to make some adjustments. This new script is able to detect the running teamviewer service and processes, stop them, and then continue with the installation.